Revert "Merge branch 'PrestaShop:develop' into patch-1"
This reverts commit 712400fe8776c7d5b8d02e25c6b128f841eb1c00, reversing
changes made to bd03107cfd421d36383f19001062fe47ba15cd93.

+++ classes/FileUploader.php
@@ -49,7 +49,7 @@ class FileUploaderCore
 
     protected function toBytes($str)
     {
-        $val = (int) trim($str);
+        $val = trim($str);
         $last = strtolower($str[strlen($str) - 1]);
         switch ($last) {
             case 'g':
